Extract shared cancellation token
