+++ tools/check-layering-paths.sh
@@ -8,6 +8,7 @@ cd "${SHELTERWOOD_ENFORCEMENT_ROOT:-.}"
 readonly below_driver_layers=(
   crates/shelterwood/src/admission.rs
   crates/shelterwood/src/actor.rs
+  crates/shelterwood/src/cancellation.rs
   crates/shelterwood/src/cells.rs
   crates/shelterwood/src/deadline.rs
   crates/shelterwood/src/definition.rs
