+++ tools/enforcement-fixtures/skeleton/crates/shelterwood/src/cancellation.rs
@@ -1,0 +1,2 @@
+// Placeholder mirroring crates/shelterwood/src/cancellation.rs so every path the
+// enforcement scripts scan exists in the fixture tree.
